DEV-3 fixed missing unit creation
